Correcting user, repo and implement pulls

+++ src/main/java/com/hacettepe/designProject/repository/OwnerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.hacettepe.designProject.entity.Owner;
 
 @Repository
-public interface OwnerRepository extends JpaRepository<Owner,Long>{
+public interface OwnerRepository extends JpaRepository<Owner,Integer>{
 
 }
